Generte python code from XML schema with pyxbgen.
This makes a load of horrible, low level files which we can
automatically parse the XML files received by the push port into. Now we
just need to wrap these files with actually user friendly objects...

The files named for.py have been renamed to fc.py in order to avoid a
conflict with python's for keyword when trying to import them.