Stub all remaining Schedule Message tests.
In order to implement these tests, we need to capture appropriate
example messages from the Push Port. They've all been added as
xfail tests for the moment so it is clear from the code what still
needs to be done to cover all the bases with tests.

Since all that reasonably can be done now is, this fixes #10.