Add schedule update test, and fix the callingpoints bug (see above)
It actually lead to a bug discovery, callingpoints weren't deleted.

NOTE: Currently it uses a test database given by the same environment
variables but with "TEST_" in front of them. If those are not given
it WILL lead to data loss. This will need to be changed in the future
for the sake of not mistakenly losing lots of data.